Add move main to proper location
